+++ src/Jwt.php
@@ -12,7 +12,6 @@ use Craft;
 use craft\base\Plugin;
 use craft\web\twig\variables\CraftVariable;
 use flipbox\craft\jwt\models\Settings as SettingsModel;
-use flipbox\craft\jwt\services\SelfConsumable;
 use Lcobucci\JWT\Builder;
 use Lcobucci\JWT\Claim\Factory as ClaimFactory;
 use Lcobucci\JWT\Parser;
@@ -27,10 +26,20 @@ use yii\log\Logger;
  * @author Flipbox Factory <hello@flipboxfactory.com>
  * @since 1.0.0
  *
+ * @property services\Identity identity
+ * @property services\Route route
+ *
  * @method SettingsModel getSettings()
  */
 class Jwt extends Plugin
 {
+    /**
+     * The plugin category
+     *
+     * @var string
+     */
+    public static $category = 'jwt';
+
     /**
      * @inheritdoc
      */
@@ -38,7 +47,13 @@ class Jwt extends Plugin
     {
         parent::init();
 
-        // Twig variables
+        // Components
+        $this->setComponents([
+            'identity' => services\Identity::class,
+            'route' => services\Route::class
+        ]);
+
+        // Template variables
         Event::on(
             CraftVariable::class,
             CraftVariable::EVENT_INIT,
@@ -63,11 +78,37 @@ class Jwt extends Plugin
      *******************************************/
 
     /**
-     * @return services\SelfConsumable
+     * @noinspection PhpDocMissingThrowsInspection
+     * @return services\Identity
      */
-    public function getSelfConsumable(): SelfConsumable
+    public function getIdentity(): services\Identity
     {
-        return $this->get('selfConsumable');
+        /** @noinspection PhpUnhandledExceptionInspection */
+        /** @noinspection PhpIncompatibleReturnTypeInspection */
+        return $this->get('identity');
+    }
+
+    /**
+     * @noinspection PhpDocMissingThrowsInspection
+     * @return services\Route
+     */
+    public function getRoute(): services\Route
+    {
+        /** @noinspection PhpUnhandledExceptionInspection */
+        /** @noinspection PhpIncompatibleReturnTypeInspection */
+        return $this->get('route');
+    }
+
+    /**
+     * @deprecated
+     */
+    public function getSelfConsumable(): services\Identity
+    {
+        Craft::$app->getDeprecator()->log(
+            self::class . '::getSelfConsumable',
+            self::class . '::getSelfConsumable() has been deprecated. Use getIdentity() instead.'
+        );
+        return $this->getIdentity();
     }
 
     /*******************************************
@@ -127,65 +168,82 @@ class Jwt extends Plugin
      *******************************************/
 
     /**
-     * Logs a trace message.
-     * Trace messages are logged mainly for development purpose to see
-     * the execution work flow of some code.
-     * @param string $message the message to be logged.
-     * @param string $category the category of the message.
+     * The log categories
+     *
+     * @param string|null $category
+     * @param bool $audit flag as an audit message.
+     * @return string
      */
-    public static function trace($message, string $category = null)
+    protected static function loggerCategory(string $category = null, bool $audit = false): string
     {
-        Craft::getLogger()->log($message, Logger::LEVEL_TRACE, self::normalizeCategory($category));
+        /** @noinspection PhpUndefinedFieldInspection */
+        $prefix = static::$category ? (static::$category . ($audit ? ':audit' : '')) : '';
+
+        if (empty($category)) {
+            return $prefix;
+        }
+
+        return ($prefix ? $prefix . ':' : '') . $category;
+    }
+
+    /**
+     * Logs a debug message.
+     * Trace messages are logged mainly for development purpose to see
+     * the execution work flow of some code. This method will only log
+     * a message when the application is in debug mode.
+     * @param string|array $message the message to be logged. This can be a simple string or a more
+     * complex data structure, such as array.
+     * @param string $category the category of the message.
+     * @param bool $audit flag as an audit message.
+     * @since 2.0.0
+     */
+    public static function debug($message, $category = 'general', bool $audit = false)
+    {
+        Craft::getLogger()->log($message, Logger::LEVEL_TRACE, static::loggerCategory($category, $audit));
     }
 
     /**
      * Logs an error message.
      * An error message is typically logged when an unrecoverable error occurs
      * during the execution of an application.
-     * @param string $message the message to be logged.
+     * @param string|array $message the message to be logged. This can be a simple string or a more
+     * complex data structure, such as array.
      * @param string $category the category of the message.
+     * @param bool $audit flag as an audit message.
+     * @since 2.0.0
      */
-    public static function error($message, string $category = null)
+    public static function error($message, $category = 'general', bool $audit = false)
     {
-        Craft::getLogger()->log($message, Logger::LEVEL_ERROR, self::normalizeCategory($category));
+        Craft::getLogger()->log($message, Logger::LEVEL_ERROR, static::loggerCategory($category, $audit));
     }
 
     /**
      * Logs a warning message.
      * A warning message is typically logged when an error occurs while the execution
      * can still continue.
-     * @param string $message the message to be logged.
+     * @param string|array $message the message to be logged. This can be a simple string or a more
+     * complex data structure, such as array.
      * @param string $category the category of the message.
+     * @param bool $audit flag as an audit message.
+     * @since 2.0.0
      */
-    public static function warning($message, string $category = null)
+    public static function warning($message, $category = 'general', bool $audit = false)
     {
-        Craft::getLogger()->log($message, Logger::LEVEL_WARNING, self::normalizeCategory($category));
+        Craft::getLogger()->log($message, Logger::LEVEL_WARNING, static::loggerCategory($category, $audit));
     }
 
     /**
      * Logs an informative message.
      * An informative message is typically logged by an application to keep record of
      * something important (e.g. an administrator logs in).
-     * @param string $message the message to be logged.
+     * @param string|array $message the message to be logged. This can be a simple string or a more
+     * complex data structure, such as array.
      * @param string $category the category of the message.
+     * @param bool $audit flag as an audit message.
+     * @since 2.0.0
      */
-    public static function info($message, string $category = null)
+    public static function info($message, $category = 'general', bool $audit = false)
     {
-        Craft::getLogger()->log($message, Logger::LEVEL_INFO, self::normalizeCategory($category));
-    }
-
-    /**
-     * @param string|null $category
-     * @return string
-     */
-    private static function normalizeCategory(string $category = null)
-    {
-        $normalizedCategory = 'JWT';
-
-        if ($category === null) {
-            return $normalizedCategory;
-        }
-
-        return $normalizedCategory . ': ' . $category;
+        Craft::getLogger()->log($message, Logger::LEVEL_INFO, static::loggerCategory($category, $audit));
     }
 }